backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
alembic==1.13.1
pydantic[email]==2.6.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
